Create a user record in the DB and user dependency.
This is for relationship mapping in the DB. I also created some tools for
injecting the user as a dependency which I will likely turn into a composite
of both the DB and the session data at some point.

Or just add the DB ID to the session. That seems faster.