Handle Ally bank uploads.
This adds fields to the Transaction row for tracking the account to/from.
This is necessary because Ally bank uploads have a description, but not
a concept of a merchant, so its hard to tie a transaction to a sourcink
directly and it can help to at least track the account.