Make the project easier to install.
I got away from this project for a while. I came back and I couldn't use my old virtual environment and had to create a new one. In doing so I simplified some of the logic for setting up the environment and documented how to get it working.
I changed a few of the dependency versions because the old ones weren't working for various reasons.
Whatever.