Add basic testing of import formats.
This is the key testing we need right now: the ability to ensure we're still
able to handle exports from different institutions. This involved a bit of a
rework of how the task code was laid out to isolate individual functions I can
reuse. Some of this work involved sorting out how to use a "test database", but
after getting that working I realized I just wanted unit tests. I'm checking in
those changes for the future.