Add support for Amex-style XLSX.
This required a really big rework to how we handle everything. We now use magic
to detect what the file seems to be, we detect the dialect of CSV, we use different
modules for the different file types, and we parse to an intermediate representation.

This is necessary because the xlsx files have multiple tabs with headers and rows within.
We therefore need to get much more sophisticated in the interface to the import processors.