Break out aspects of app setup.
The point here is that we need to create a testing environment where the app
is setup the same way but our database is totally different. It makes sense
then to have a separate entrypoint for uvicorn, our prod app and DB, and
the main appilcation configuration. Test will then have its own entrypoint
and use the main application configuration to ensure the app is as similar
as can be for tests.

I've broken out the steps of DB setup using chained Depends and we'll
eventually override the get_config for testing so that the DB connection
gets a different config.