Add category inference to process page.
This is a bit sloppy in several ways, but the basics are present and work.
I wasn't sure if the training would reliably be fast enough to to at request
time so I kick off a background task and save the data locally in the current
directory. This is pretty racy. Additionally I'm just serializing to a raw
tuple which is ugly and will likely create bugs later.

But it works, it correctly infers, which is really cool.

There are likely some gains to be had by separating out by account type. I also
need to work on some logic for detecting when we are not or should not be confident
about the suggestion so we avoid it.

Baby steps.